Quick and dirty work around to address Issue #2
Essentially I'm just checking if  "bufferIndex" is out of range; if it is setting it to a value that will skip processing of that packet. This will make the program not crash but it may also not function as intended either! A real fix will probably require more coding and I just haven't had the time lately.